All tests passed.

+++ src/route_planner.cpp
@@ -63,8 +63,8 @@ void RoutePlanner::AddNeighbors(RouteModel::Node *current_node) {
 
 // Helper Function - Compares nodes a and b f_values
 bool Compare(const RouteModel::Node* node_a, RouteModel::Node* node_b) {
-    int f_node_a = node_a->g_value + node_a->h_value; // fa = ga + ha
-    int f_node_b = node_b->g_value + node_b->h_value; // fb = gb + hb
+    float f_node_a = node_a->g_value + node_a->h_value; // fa = ga + ha
+    float f_node_b = node_b->g_value + node_b->h_value; // fb = gb + hb
     return f_node_a > f_node_b; 
 }
